HR.LeaveManagement.Application/Features/LeaveType/Queries/GetLeaveTypeDetails/GetLeaveTypeDetailsQueryHandler.cs
using AutoMapper;
using HR.LeaveManagement.Application.Contracts.Persistence;
using HR.LeaveManagement.Application.Exceptions;
using HR.LeaveManagement.Domain;
using MediatR;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HR.LeaveManagement.Application.Features.LeaveType.Queries.GetLeaveTypeDetails
{
    public class GetLeaveTypeDetailsQueryHandler : IRequestHandler<GetLeaveTypeDetailsQuery, LeaveTypeDetailsDto>
    {
        private readonly ILeaveTypeRepository leaveTypeRepository;
        private readonly IMapper mapper;

        public GetLeaveTypeDetailsQueryHandler(ILeaveTypeRepository leaveTypeRepository, IMapper mapper)
        {
            this.leaveTypeRepository = leaveTypeRepository;
            this.mapper = mapper;
        }
        public async Task<LeaveTypeDetailsDto> Handle(GetLeaveTypeDetailsQuery request, CancellationToken cancellationToken)
        {
            //Query LeaveType 
            var LeaveType= await leaveTypeRepository.GetByIdAsync(request.Id);
            //verify record exists
            if (LeaveType is null)
                throw new NotFoundException(nameof(LeaveType), request.Id);
            //AutoMap 
            var data = mapper.Map<LeaveTypeDetailsDto>(LeaveType);
            //return the leave type details
            return data;
        }
    }
}
